+++ x-pack/plugins/cases/server/common/types/configure.ts
@@ -8,12 +8,12 @@
 import * as rt from 'io-ts';
 
 import type { SavedObject } from '@kbn/core/server';
-import type { ConfigurationAttributes } from '../../../common/api';
+import type { ConfigurationAttributes } from '../../../common/types/domain';
 import {
   ConfigurationActivityFieldsRt,
-  ConfigurationBasicWithoutOwnerRt,
   ConfigurationAttributesRt,
-} from '../../../common/api';
+  ConfigurationBasicWithoutOwnerRt,
+} from '../../../common/types/domain';
 import type { ConnectorPersisted } from './connectors';
 import type { User } from './user';
