+++ x-pack/plugins/cases/server/services/configure/index.test.ts
@@ -5,11 +5,7 @@
  * 2.0.
  */
 
-import type {
-  CaseConnector,
-  ConfigurationAttributes,
-  ConfigurationPatchRequest,
-} from '../../../common/api';
+import type { CaseConnector } from '../../../common/api';
 import { ConnectorTypes } from '../../../common/api';
 import { CASE_CONFIGURE_SAVED_OBJECT, SECURITY_SOLUTION_OWNER } from '../../../common/constants';
 import { savedObjectsClientMock } from '@kbn/core/server/mocks';
@@ -30,6 +26,8 @@ import type { ESCaseConnectorWithId } from '../test_utils';
 import { createESJiraConnector, createJiraConnector } from '../test_utils';
 import type { ConfigurationPersistedAttributes } from '../../common/types/configure';
 import { unset } from 'lodash';
+import type { ConfigurationPatchRequest } from '../../../common/types/api';
+import type { ConfigurationAttributes } from '../../../common/types/domain';
 
 const basicConfigFields = {
   closure_type: 'close-by-pushing' as const,
