+++ x-pack/plugins/cases/server/services/configure/types.ts
@@ -6,7 +6,7 @@
  */
 
 import type { SavedObject, SavedObjectsClientContract } from '@kbn/core/server';
-import type { ConfigurationAttributes } from '../../../common/api';
+import type { ConfigurationAttributes } from '../../../common/types/domain';
 import type { IndexRefresh } from '../types';
 import type { SavedObjectFindOptionsKueryNode } from '../../common/types';
